Move sandbox function mutation hooks to SWR
