+++ front/components/actions/blocked/SandboxFunctionPersonalAuthCard.tsx
@@ -1,8 +1,8 @@
 import type { PersonalAuthResolutionOutcome } from "@app/components/actions/blocked/PersonalAuthenticationCard";
 import { PersonalAuthenticationCard } from "@app/components/actions/blocked/PersonalAuthenticationCard";
-import { useResolveSandboxFunctionAuthentication } from "@app/hooks/useResolveSandboxFunctionAuthentication";
 import type { SandboxFunctionToolPersonalAuthRequiredEvent } from "@app/lib/actions/mcp_internal_actions/events";
 import { useAuth } from "@app/lib/auth/AuthContext";
+import { useResolveSandboxFunctionAuthentication } from "@app/lib/swr/sandbox_functions";
 
 interface SandboxFunctionPersonalAuthCardProps {
   event: SandboxFunctionToolPersonalAuthRequiredEvent;
