+++ front/components/actions/blocked/SandboxFunctionToolApprovalCard.tsx
@@ -1,8 +1,8 @@
 import { ToolValidationCard } from "@app/components/actions/blocked/ToolValidationCard";
-import { useValidateSandboxFunctionAction } from "@app/hooks/useValidateSandboxFunctionAction";
 import type { MCPValidationOutputType } from "@app/lib/actions/constants";
 import type { SandboxFunctionMCPApproveExecutionEvent } from "@app/lib/actions/mcp_internal_actions/events";
 import { useAuth } from "@app/lib/auth/AuthContext";
+import { useValidateSandboxFunctionAction } from "@app/lib/swr/sandbox_functions";
 import { useState } from "react";
 
 interface SandboxFunctionToolApprovalCardProps {
